samples/callbacks/function/requirements.txt
google-cloud-firestore>=2.0.1